🐛fix: agent memory lost due to the base64 image
