+++ packages/mcp-android-server/src/tools/ai-action.ts
@@ -2,6 +2,7 @@ import { z } from "zod";
 import { attachBox } from "../gboxsdk/index.js";
 import type { MCPLogger } from "../mcp-logger.js";
 import type { ActionAI } from "gbox-sdk";
+import { sanitizeResult } from "./utils.js";
 
 
 export const AI_ACTION_TOOL = "ai_action";
@@ -93,40 +94,6 @@ export function handleAiAction(logger: MCPLogger) {
 
       // Build content array with text and images
       const content: Array<{ type: "text"; text: string } | { type: "image"; data: string; mimeType: string }> = [];
-
-      // Create a sanitized version of result with truncated base64 data
-      const sanitizeResult = (obj: any): any => {
-        if (typeof obj !== 'object' || obj === null) {
-          return obj;
-        }
-        
-        if (Array.isArray(obj)) {
-          return obj.map(sanitizeResult);
-        }
-        
-        const sanitized: any = {};
-        for (const [key, value] of Object.entries(obj)) {
-          if (key === 'uri' && typeof value === 'string') {
-            // Truncate base64 data URIs
-            if (value.startsWith('data:')) {
-              const match = value.match(/^(data:.+;base64,)(.*)$/);
-              if (match && match[2].length > 20) {
-                sanitized[key] = match[1] + match[2].substring(0, 20) + '...';
-              } else {
-                sanitized[key] = value;
-              }
-            } else if (value.length > 20) {
-              // Truncate other long strings that might be base64
-              sanitized[key] = value.substring(0, 20) + '...';
-            } else {
-              sanitized[key] = value;
-            }
-          } else {
-            sanitized[key] = sanitizeResult(value);
-          }
-        }
-        return sanitized;
-      };
 
       // Add text result with sanitized data
       content.push({
